# packages/contracts/openapi/nexweave-platform-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "extensions": {
            "additionalProperties": true,
            "title": "Extensions",
            "type": "object"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "WorkflowTaskCreate": {
        "additionalProperties": false,
        "properties": {
          "business_key": {
            "maxLength": 255,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Business Key",
            "type": "string"
          },
          "display_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "input_refs": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Input Refs",
            "type": "object"
          },
          "start_paused": {
            "default": false,
            "title": "Start Paused",
            "type": "boolean"
          },
          "workflow_type": {
            "$ref": "#/components/schemas/WorkflowType"
          }
        },
        "required": [
          "workflow_type",
          "business_key",
          "display_name"
        ],
        "title": "WorkflowTaskCreate",
        "type": "object"
      },
      "WorkflowTaskDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/WorkflowCommand"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "task": {
            "$ref": "#/components/schemas/WorkflowTaskResponse"
          }
        },
        "required": [
          "task",
          "steps",
          "events",
          "allowed_actions"
        ],
        "title": "WorkflowTaskDetailResponse",
        "type": "object"
      },
      "WorkflowTaskListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowTaskResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "WorkflowTaskListResponse",
        "type": "object"
      },
...
      "WorkflowTaskStatus": {
        "enum": [
          "CREATED",
          "STARTING",
          "RUNNING",
          "PAUSED",
          "WAITING",
          "WAITING_INPUT",
          "CANCELLING",
          "COMPENSATING",
          "CANCELLED",
          "SUCCEEDED",
          "FAILED",
          "TIMED_OUT",
          "REJECTED"
        ],
        "title": "WorkflowTaskStatus",
        "type": "string"
      },
      "WorkflowType": {
        "enum": [
          "SOURCE_INGESTION",
          "KNOWLEDGE_COMPILE",
          "HUMAN_REVIEW",
          "QUALITY_EVALUATION",
          "KNOWLEDGE_RELEASE",
          "DOMAIN_PACK_INSTALL",
          "GRIDCREW_FEEDBACK_INGESTION"
        ],
        "title": "WorkflowType",
        "type": "string"
      }
    },
    "securitySchemes": {
      "OIDC Bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "M2 reliable Temporal workflow kernel and real task-center APIs.",
    "title": "NEXWEAVE Platform API",
    "version": "1.2.0-m2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/spaces/{space_id}/workflow-tasks": {
      "get": {
        "operationId": "list_workflow_tasks_api_v1_spaces__space_id__workflow_tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "workflow_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkflowType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkflowTaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          }
        },
        "security": [
          {
            "OIDC Bearer": []
          }
        ],
        "summary": "List Workflow Tasks",
        "tags": [
          "workflow-tasks"
        ]
      },
      "post": {
        "operationId": "create_workflow_task_api_v1_spaces__space_id__workflow_tasks_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowTaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          }
        },
        "security": [
          {
            "OIDC Bearer": []
          }
        ],
        "summary": "Create Workflow Task",
        "tags": [
          "workflow-tasks"
        ]
      }
    },
...
    "/api/v1/workflow-tasks/{task_id}": {
      "get": {
        "operationId": "get_workflow_task_api_v1_workflow_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowTaskDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          }
        },
        "security": [
          {
            "OIDC Bearer": []
          }
        ],
        "summary": "Get Workflow Task",
        "tags": [
          "workflow-tasks"
        ]
      }
    },
    "/api/v1/workflow-tasks/{task_id}/commands": {
      "post": {
        "operationId": "command_workflow_task_api_v1_workflow_tasks__task_id__commands_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          }
        },
        "security": [
          {
            "OIDC Bearer": []
          }
        ],
        "summary": "Command Workflow Task",
        "tags": [
          "workflow-tasks"
        ]
      }
    },
    "/api/v1/workflow-tasks/{task_id}/reconcile": {
      "post": {
        "operationId": "reconcile_workflow_task_api_v1_workflow_tasks__task_id__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowReconcileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details"
          }
        },
        "security": [
          {
            "OIDC Bearer": []
          }
        ],
        "summary": "Reconcile Workflow Task",
        "tags": [
          "workflow-tasks"
        ]
      }
    }
  }
}